+++ chillz/departments.json
@@ -1,0 +1,141 @@
+
+
+{
+  "departments": [
+    {
+      "code": "AE"
+    },
+    {
+      "code": "AG"
+    },
+    {
+      "code": "AR"
+    },
+    {
+      "code": "AT"
+    },
+    {
+      "code": "BE"
+    },
+    {
+      "code": "BM"
+    },
+    {
+      "code": "BS"
+    },
+    {
+      "code": "BT"
+    },
+    {
+      "code": "CD"
+    },
+    {
+      "code": "CE"
+    },
+    {
+      "code": "CH"
+    },
+    {
+      "code": "CL"
+    },
+    {
+      "code": "CR"
+    },
+    {
+      "code": "CS"
+    },
+    {
+      "code": "CY"
+    },
+    {
+      "code": "DE"
+    },
+    {
+      "code": "EC"
+    },
+    {
+      "code": "EE"
+    },
+    {
+      "code": "EF"
+    },
+    {
+      "code": "ES"
+    },
+    {
+      "code": "ET"
+    },
+    {
+      "code": "GG"
+    },
+    {
+      "code": "GS"
+    },
+    {
+      "code": "HS"
+    },
+    {
+      "code": "ID"
+    },
+    {
+      "code": "IM"
+    },
+    {
+      "code": "IP"
+    },
+    {
+      "code": "IT"
+    },
+    {
+      "code": "MA"
+    },
+    {
+      "code": "ME"
+    },
+    {
+      "code": "MI"
+    },
+    {
+      "code": "MM"
+    },
+    {
+      "code": "MS"
+    },
+    {
+      "code": "MT"
+    },
+    {
+      "code": "NA"
+    },
+    {
+      "code": "NT"
+    },
+    {
+      "code": "PH"
+    },
+    {
+      "code": "RD"
+    },
+    {
+      "code": "RE"
+    },
+    {
+      "code": "RJ"
+    },
+    {
+      "code": "RT"
+    },
+    {
+      "code": "RX"
+    },
+    {
+      "code": "SL"
+    },
+    {
+      "code": "TS"
+    },
+    {
+      "code": "WM"
+    }
+  ]
+}
